search: RepositoryRevSpecs returns an iterator
This can be a very large list and the underlying repos resolver relies
on an iterator to avoid the large list appearing in memory. So this
commit adjusts the interface to also use an iterator reducing memory
usage the first step of exhaustive search.

Test Plan: go test should be sufficient